+++ NEWS.md
@@ -1,5 +1,9 @@
 ## qfmtr 0.0.0.9000
 
+To do:
+
+- adds function `qfmt()` to format a list to be used in a query
+
 - adds **pkgdown** support
 - adds a `NEWS.md` file to track changes to the package